정답률 추이 값을 정수로 반올림
- AccuracyTrendResponse.accuracy를 double에서 int로 변경하고 Math.round 적용
- 대시보드 평균 정답률이 "33.33333..%" 대신 정수로 표시되도록 수정

+++ backend/src/main/java/com/pokemo/stats/api/AccuracyTrendResponse.java
@@ -2,5 +2,5 @@ package com.pokemo.stats.api;
 
 public record AccuracyTrendResponse(
     String attemptedAt,
-    double accuracy
+    int accuracy
 ) {}
